+++ src/main/scala/com/sfxcode/music/itunes/model/Playlist.scala
@@ -8,9 +8,23 @@ import scala.collection.mutable
 
 case class Playlist(data: PlaylistData, tracks: List[Track], customValues: mutable.HashMap[String, Any] = new mutable.HashMap()) {
 
+  var parent:Option[Playlist] = None
+  var children:List[Playlist] = List()
+
   def id: Long = data.id
+  def persistantKey:String  = data.persistantKey
+  def parentPersistantKey: String = data.parentPersistantKey
 
   def name: String = data.name
+
+  def parentName = {
+    if (parent.isDefined)
+      parent.get.name
+    else
+      PlaylistData.EmptyParentString
+  }
+
+
 
   def kind: Int = data.kind
 
@@ -19,6 +33,8 @@ case class Playlist(data: PlaylistData, tracks: List[Track], customValues: mutab
   def music: Boolean = data.music
 
   def master: Boolean = data.master
+
+  def folder: Boolean = data.folder
 
   def totalTime: Int = tracks.map(track => track.totalTime).sum